Fix build warning on 64 bit machines in machdep.h (used long instead of int for ptr casting for calculus)
git-svn-id: file:///home/jj/hercules.svn/trunk@3141 956126f8-22a0-4046-8f4a-272fa8102e63

+++ machdep.h
@@ -709,7 +709,7 @@ static __inline__ void concpy ( void *_dest, void *_src, size_t n )
     }
 
     /* copy to dest double-word boundary */
-    n2 = 8 - ((int)dest & 7);
+    n2 = 8 - ((long)dest & 7);
     if (n2 < 8)
     {
         n -= n2;
@@ -721,7 +721,7 @@ static __inline__ void concpy ( void *_dest, void *_src, size_t n )
     {
 #if defined(ASSIST_CMPXCHG8)
         /* copy unaligned double-words */
-        if ((int)src & 7)
+        if ((long)src & 7)
             do {
                 U64 temp;
                 memcpy(&temp, src, 8);
